refactor: SetPlace.tsx component refactoring
