+++ src/pages/createMeeting/components/ReturnBodyComponent.tsx
@@ -12,6 +12,7 @@ import { theme } from 'styles/theme';
 
 import SetTitle from './useFunnel/SetTitle';
 import { MeetingInfo } from '../types/useFunnelInterface';
+import SetPlace from './useFunnel/SetPlace';
 
 interface BodyProps {
   currentStep: string;
@@ -27,10 +28,6 @@ type BodyType = {
 function ReturnBodyComponent({ currentStep, meetingInfo, setMeetingInfo, setStep }: BodyProps) {
 
   console.log(meetingInfo)
-
-  const setPlaceDetail = (place: string) => {
-    setMeetingInfo((prev) => ({ ...prev, place ,placeDetail: ""}));
-  };
 
   const passWordOnChange = (e: React.ChangeEvent<HTMLInputElement>) => {
     setMeetingInfo((prev : MeetingInfo) => {
@@ -111,40 +108,7 @@ function ReturnBodyComponent({ currentStep, meetingInfo, setMeetingInfo, setStep
     ),
     place: (
       <>
-        <PlaceInfoWrapper>
-          {placeType.map((type , i)=>{
-            return (<PlaceSetion key={i+type}>
-                      <Button typeState={meetingInfo?.place === type ? 'primaryActive' : 'secondaryDisabled'} onClick={()=>setPlaceDetail(type)}>
-                          <Text font={'button2'}>{type === "ONLINE"?"온라인": type ==="OFFLINE"?"오프라인":"미정"}</Text>
-                      </Button>
-                      {type === "UNDEFIND" ? null : meetingInfo?.place === type ? <PlaceInput
-                        data={type}
-                        value={meetingInfo.placeDetail}
-                        setValue={setMeetingInfo}
-                        placeholder={type === "ONLINE"?"(선택) 화상 회의 툴을 입력해주세요":"(선택) 구체적인 장소명을 입력해주세요"}
-                      /> : null
-                      }
-                  </PlaceSetion>)
-          })}
-        </PlaceInfoWrapper>
-        <StyledBtnWrapper>
-          <Button
-            typeState={meetingInfo.place ? 'primaryActive' : 'secondaryDisabled'}
-            onClick={
-                meetingInfo.place
-                ? () =>
-                    setStep((prev) => {
-                      if (prev === 6) {
-                        return prev;
-                      }
-                      return prev + 1;
-                    })
-                : undefined
-            }
-          >
-            <Text font={'button2'}>다음</Text>
-          </Button>
-        </StyledBtnWrapper>
+        <SetPlace meetingInfo={meetingInfo} setMeetingInfo={setMeetingInfo} setStep={setStep} />
       </>
     ),
     duration: (
